+++ libraries/testable/testcase.php
@@ -38,6 +38,7 @@ abstract class TestCase extends PHPUnit_Framework_TestCase {
 		{
 			Event::listen('orchestra.testable: setup-db', function ()
 			{
+				File::delete(path('storage').'database'.DS.'orchestra.sqlite');
 				Config::set('database.connections.testdb', array(
 					'driver'   => 'sqlite',
 					'database' => 'orchestra',